Size the SHA-1 digest at its defined width

+++ code/sha.h
@@ -32,6 +32,7 @@
 #pragma once
 
 
+#include <cstdint>
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
@@ -68,7 +69,7 @@ class SHAEngine
 	private:
 
 		typedef union {
-			unsigned long Long[5];
+			uint32_t Long[5];
 			unsigned char Char[20];
 		} SHADigest;
